+++ services/DBService/src/main/kotlin/boysband/dbservice/repository/UserRepository.kt
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository
 
 @Repository
 interface UserRepository : JpaRepository<User, Int> {
-    fun findByIdTgChat(idTgChat: Int): User?
+    fun findByIdTgChat(idTgChat: Long): User?
 }
